refactor: period 조건 통해 광고 개수 조회 - 날짜/차수 조건에 따른 매체명/업종명 광고수 조회하기 위해 period 파라미터 추가 - AdvertisementRepositoryCustom에서 해당하는 차수에 각 매체명/업종명 비교하여 광고수 조회

+++ src/main/java/com/fastcampus/befinal/domain/info/FilterInfo.java
@@ -6,9 +6,9 @@ public class FilterInfo {
     @Builder
     public record FilterOptionInfo(
         String name,
-        int adCount
+        Long adCount
     ) {
-        public static FilterOptionInfo of(String name, int adCount) {
+        public static FilterOptionInfo of(String name, Long adCount) {
             return FilterOptionInfo.builder()
                 .name(name)
                 .adCount(adCount)
